# docs/openapi/app.openapi.yaml
openapi: 3.0.3
info:
  title: 鐪熷ソ鐜╁皬绋嬪簭绔?API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/app/connections/{connectionId}/follow-logs:
    post:
      summary: Create connection follow log
      security:
        - bearerAuth: []
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Creates a follow log for the connection owner; approved guides can also create guide follow logs.
        "403":
          description: Current user cannot follow this connection.
        "404":
          description: Connection not found.
  /api/app/connections/{connectionId}/follow-up:
    post:
      summary: Create connection follow log compatibility alias
      security:
        - bearerAuth: []
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Compatibility alias for /api/app/connections/{connectionId}/follow-logs.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer